extract_utils: Implement overriding of elf format
Allow for manual overriding of the ELF format, which can be useful in cases
where a blob's format cannot be read due to an invalid magic number.

It is also beneficial in cases where a blob is 32-bit but located in lib64/ while the build is 64-bit.
In such cases, the build system attempts to compile a 64-bit version of the blob.

$ llvm-objdump -a odm/lib64/libAncHumanRetain.so
> odm/lib64/libAncHumanRetain.so: file format elf32-littlearm

Change-Id: I3bd30e2f61f7e000e677540e7682b5fbcd1e49c8